fix(web): stabilize Den onboarding flow with app routes
