+++ packages/web/app/(den)/checkout/page.tsx
@@ -1,0 +1,9 @@
+import { CheckoutScreen } from "../_components/checkout-screen";
+
+export default function CheckoutPage({
+  searchParams,
+}: {
+  searchParams?: { customer_session_token?: string };
+}) {
+  return <CheckoutScreen customerSessionToken={searchParams?.customer_session_token ?? null} />;
+}
