+++ packages/web/app/(den)/dashboard/page.tsx
@@ -1,0 +1,5 @@
+import { DashboardScreen } from "../_components/dashboard-screen";
+
+export default function DashboardPage() {
+  return <DashboardScreen />;
+}
