+++ packages/web/app/(den)/layout.tsx
@@ -1,0 +1,11 @@
+import type { ReactNode } from "react";
+import { DenShell } from "./_components/den-shell";
+import { DenFlowProvider } from "./_providers/den-flow-provider";
+
+export default function DenLayout({ children }: { children: ReactNode }) {
+  return (
+    <DenFlowProvider>
+      <DenShell>{children}</DenShell>
+    </DenFlowProvider>
+  );
+}
