+++ packages/web/app/(den)/page.tsx
@@ -1,0 +1,5 @@
+import { AuthScreen } from "./_components/auth-screen";
+
+export default function HomePage() {
+  return <AuthScreen />;
+}
